tests/local/tree-construction/remex01.dat

#data
<pre><!doctype html>
x</pre>
#errors
[0] missing doctype
[5] unexpected doctype
#document
| <html>
|   <head>
|   <body>
|     <pre>
|       "
x"
